refactor: scope and bound the flag-field interner
Move the string interner out of the global util barrel into the
client-feature-toggles module, rename internString -> internFlagField, and
cap the pool at 1000 entries. A global unopinionated internString invited
interning high-cardinality values (user ids, descriptions), which would grow
the pool without bound. Scoping it, naming it for its purpose, and bounding
the pool make misuse safe: past the cap it stops pooling and just returns the
value.
