+++ src/lib/features/client-feature-toggles/delta/client-feature-toggle-delta-read-model.ts
@@ -6,9 +6,9 @@ import type EventEmitter from 'events';
 import {
     ALL_PROJECTS,
     ensureStringValue,
-    internString,
     mapValues,
 } from '../../../util/index.js';
+import { internFlagField } from '../intern-flag-field.js';
 import type {
     FeatureConfigurationDeltaClient,
     IClientFeatureToggleDeltaReadModel,
@@ -167,7 +167,7 @@ export default class ClientFeatureToggleDeltaReadModel
             feature.description = r.description;
             feature.project = r.project;
             feature.stale = r.stale;
-            feature.type = internString(r.type);
+            feature.type = internFlagField(r.type);
             feature.variants = r.variants || [];
             feature.project = r.project;
 
@@ -217,7 +217,7 @@ export default class ClientFeatureToggleDeltaReadModel
     private rowToStrategy(row: Record<string, any>): IStrategyConfig {
         const strategy: IStrategyConfig = {
             id: row.strategy_id,
-            name: internString(row.strategy_name),
+            name: internFlagField(row.strategy_name),
             title: row.strategy_title,
             constraints: row.constraints || [],
             parameters: mapValues(row.parameters || {}, ensureStringValue),
