+++ src/lib/features/client-feature-toggles/intern-flag-field.test.ts
@@ -1,0 +1,7 @@
+import { createFlagFieldInterner } from './intern-flag-field.js';
+
+test('returns a value equal to the input', () => {
+    const internFlagField = createFlagFieldInterner();
+
+    expect(internFlagField('flexibleRollout')).toBe('flexibleRollout');
+});
