+++ src/lib/features/frontend-api/client-feature-toggle-read-model.ts
@@ -4,11 +4,8 @@ import type {
     IStrategyConfig,
     PartialDeep,
 } from '../../types/index.js';
-import {
-    ensureStringValue,
-    internString,
-    mapValues,
-} from '../../util/index.js';
+import { ensureStringValue, mapValues } from '../../util/index.js';
+import { internFlagField } from '../client-feature-toggles/intern-flag-field.js';
 import type { Db } from '../../db/db.js';
 import FeatureToggleStore from '../feature-toggle/feature-toggle-store.js';
 type Raw<T = any> = Knex.Raw<T>;
@@ -125,7 +122,7 @@ export default class ClientFeatureToggleReadModel
                     description: row.description,
                     project: row.project,
                     stale: row.stale,
-                    type: internString(row.type),
+                    type: internFlagField(row.type),
                 };
             }
 
@@ -187,7 +184,7 @@ export default class ClientFeatureToggleReadModel
     private rowToStrategy(row: Record<string, any>): IStrategyConfig {
         const strategy: IStrategyConfig = {
             id: row.strategy_id,
-            name: internString(row.strategy_name),
+            name: internFlagField(row.strategy_name),
             title: row.strategy_title,
             constraints: row.constraints || [],
             parameters: mapValues(row.parameters || {}, ensureStringValue),
