--- src/lib/util/intern-string.test.ts
@@ -1,5 +1,0 @@
-import { internString } from './intern-string.js';
-
-test('returns a value equal to the input', () => {
-    expect(internString('flexibleRollout')).toBe('flexibleRollout');
-});
